pyproject.toml
[tool.poetry]
name = "volta-cli"
version = "0.1.0"
description = ""
authors = ["zani-t <tembo4.t@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
volta-cli = "volta_cli.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.6.1"}
Flask = "^2.2.2"
mysql-connector-python = "^8.0.31"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
